pyproject.toml
[tool.poetry]
name = "wz-bot"
version = "0.1.0"
description = ""
authors = ["tagliani97 <gabrieltagliani30@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
bs4 = "^0.0.1"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
